[r=allenap][bug=][author=mpontillo] Add extended backward compatibility for Juju.
 * Return the undefined space in the API as if it is a space.
 * Supports reading the undefined space as a single space, or as
   a part of the list of spaces, but only when VLANs exist which are
   not associated with a space.

+++ src/maasserver/api/spaces.py
@@ -3,16 +3,21 @@
 
 """API handlers: `Space`."""
 
+from django.db.models.query import QuerySet
 from maasserver.api.support import (
     admin_method,
     OperationsHandler,
 )
 from maasserver.enum import NODE_PERMISSION
-from maasserver.exceptions import MAASAPIValidationError
+from maasserver.exceptions import (
+    MAASAPIBadRequest,
+    MAASAPIValidationError,
+)
 from maasserver.forms_space import SpaceForm
 from maasserver.models import (
     Space,
     Subnet,
+    VLAN,
 )
 from piston3.utils import rc
 
@@ -24,6 +29,31 @@ DISPLAYED_SPACE_FIELDS = (
     'vlans',
     'subnets',
 )
+
+
+def _has_undefined_space():
+    """Returns True if the undefined space contains at least one VLAN."""
+    return VLAN.objects.filter(space__isnull=True).exists()
+
+
+# Placeholder Space-like object for backward compatibility.
+UNDEFINED_SPACE = Space(
+    id=-1, name=Space.UNDEFINED,
+    description="Backward compatibility object to ensure objects not "
+                "associated with a space can be found.")
+
+UNDEFINED_SPACE.save = None
+
+
+class SpacesQuerySet(QuerySet):
+
+    def __iter__(self):
+        """Custom iterator which also includes a dummy "undefined" space."""
+        yield from super().__iter__()
+        # This space will be related to any VLANs and subnets not associated
+        # with a space. (For backward compatibility with Juju 2.0.)
+        if _has_undefined_space():
+            yield UNDEFINED_SPACE
 
 
 class SpacesHandler(OperationsHandler):
@@ -39,7 +69,12 @@ class SpacesHandler(OperationsHandler):
 
     def read(self, request):
         """List all spaces."""
-        return Space.objects.all()
+        spaces_query = Space.objects.all()
+        # The .all() method will return a QuerySet, but we need to coerce it to
+        # a SpacesQuerySet to get our custom iterator. This must be an instance
+        # of a QuerySet, or the API framework will return it as-is.
+        spaces_query.__class__ = SpacesQuerySet
+        return spaces_query
 
     @admin_method
     def create(self, request):
@@ -67,7 +102,10 @@ class SpaceHandler(OperationsHandler):
         # See the comment in NodeHandler.resource_uri.
         space_id = "id"
         if space is not None:
-            space_id = space.id
+            if space.id == -1:
+                space_id = Space.UNDEFINED
+            else:
+                space_id = space.id
         return ('space_handler', (space_id,))
 
     @classmethod
@@ -79,10 +117,16 @@ class SpaceHandler(OperationsHandler):
 
     @classmethod
     def subnets(cls, space):
+        # Backward compatibility for Juju 2.0.
+        if space.id == -1:
+                return Subnet.objects.filter(vlan__space__isnull=True)
         return Subnet.objects.filter(vlan__space=space)
 
     @classmethod
     def vlans(cls, space):
+        # Backward compatibility for Juju 2.0.
+        if space.id == -1:
+            return VLAN.objects.filter(space__isnull=True)
         return space.vlan_set.all()
 
     def read(self, request, id):
@@ -90,6 +134,10 @@ class SpaceHandler(OperationsHandler):
 
         Returns 404 if the space is not found.
         """
+        # Backward compatibility for Juju 2.0. This is a special case to check
+        # if the user requested to read the undefined space.
+        if id == "-1" or id == Space.UNDEFINED and _has_undefined_space():
+            return UNDEFINED_SPACE
         return Space.objects.get_space_or_404(
             id, request.user, NODE_PERMISSION.VIEW)
 
@@ -101,6 +149,9 @@ class SpaceHandler(OperationsHandler):
 
         Returns 404 if the space is not found.
         """
+        if id == "-1" or id == Space.UNDEFINED:
+            raise MAASAPIBadRequest(
+                "Space cannot be modified: %s" % Space.UNDEFINED)
         space = Space.objects.get_space_or_404(
             id, request.user, NODE_PERMISSION.ADMIN)
         form = SpaceForm(instance=space, data=request.data)
@@ -114,6 +165,9 @@ class SpaceHandler(OperationsHandler):
 
         Returns 404 if the space is not found.
         """
+        if id == "-1" or id == Space.UNDEFINED:
+            raise MAASAPIBadRequest(
+                "Space cannot be deleted: %s" % Space.UNDEFINED)
         space = Space.objects.get_space_or_404(
             id, request.user, NODE_PERMISSION.ADMIN)
         space.delete()
